requirements.txt
Flask==2.1.1
Flask-RESTful==0.3.9
psycopg2-binary==2.9.3
